lib/: X(): Add macro to exit() on error
Writing an x*() variant of several functions is unnecessary.  It's
simpler to write a generic X() macro that can be chained with any other
calls.

For example:

	-  X(malloc(...)) instead of xmalloc(...);
	-  X(strdup(...)) instead of xstrdup(...);

If an error is detected, log an error, and exit(13).  About why 13, I
don't really know.  It's just what was used previously in xmalloc().

Signed-off-by: Alejandro Colomar <alx@kernel.org>