lib/alloc/realloc*.h: Always reallocate at least 1 byte
glibc's realloc(3) is broken.  It was good until 2006/2007, when it was
changed to conform to C89, which had a bogus specification that required
that it returns NULL.  However, by the time glibc was changed to
conform to the broken C89, the standard specification had already been
fixed in C99.  Thus, the change in glibc made glibc non-conforming to
C99 (it had been conforming previously).

Link: <https://github.com/shadow-maint/shadow/pull/1095>
Link: <https://nabijaczleweli.xyz/content/blogn_t/017-malloc0.html>
Co-developed-by: наб <nabijaczleweli@nabijaczleweli.xyz>
Signed-off-by: наб <nabijaczleweli@nabijaczleweli.xyz>
Signed-off-by: Alejandro Colomar <alx@kernel.org>