lib/, src/: Use default argument values in a2i() calls
This removes chances of typos.  It would be possible to accidentally
pass for example ULONG_MAX or INT_MAX where LONG_MAX would be
appropriate.  Because these macros are able to automate the right value,
don't specify it manually.

Thus, when we specify it manually, it will be in those places where we
use a non-default value, which is less common.  That will make it easier
to review those few places.

Signed-off-by: Alejandro Colomar <alx@kernel.org>