src/newgrp.c: Refactor conditionals to de-duplicate code
Both sg(1) and newgrp(1) handle the group name in the same way if it
exists.  They differ in how they handle the absence of it.  Reflect that
in the source code.

Use a block to avoid changing the indentation.  This will make the diff
easier to review.  A future patch will clean up the indentation.

Signed-off-by: Alejandro Colomar <alx@kernel.org>